add postgresql.conf from outside into the database container
We are starting to control-version the changes of the
postgres config file. Notice that this initial file has been
extracted from the current store-db-01.do-ams3.shards.staging node.

I only added some parameters, commented, that will be useful for us
to debug the database. To find those new parameters just look
for the 'waku' word within the postgresql.conf file.